Switched to float

+++ c++/camera.hpp
@@ -9,17 +9,17 @@ class camera {
         vec3 horizontal;
         vec3 vertical;
         vec3 u, v, w;
-        double lens_radius;
+        float lens_radius;
 
     public:
         camera(
             point3 lookfrom,
             point3 lookat,
             vec3   vup,
-            double vfov, // vertical field-of-view in degrees
-            double aspect_ratio,
-            double aperture,
-            double focus_dist
+            float vfov, // vertical field-of-view in degrees
+            float aspect_ratio,
+            float aperture,
+            float focus_dist
         ) {
             auto theta = degrees_to_radians(vfov);
             auto h = tan(theta/2);
@@ -39,7 +39,7 @@ class camera {
         }
 
 
-        ray get_ray(double s, double t) const {
+        ray get_ray(float s, float t) const {
             vec3 rd = lens_radius * uniform_random_in_unit_disk();
             vec3 offset = u * rd.x() + v * rd.y();
