+++ c++/ray.hpp
@@ -13,7 +13,7 @@ class ray{
         point3 origin() const {return orig;}
         vec3 direction() const {return dir;}
 
-        point3 at(double t) const {
+        point3 at(float t) const {
             return orig + t * dir;
         }
 };
